Allow a `fetch` function to be passed into apiMiddleware
This was done in the most API-preserving way possible. Another (breaking)
change would be to force the library user to initialize the apiMiddleware
with the given fetch. I'd actually prefer that, but thought that preserving
the existing API would be the most important priority right now.

Note also that this way of doing it is slightly leaky as passing in fetch
essentially modifies a global-ish variable. The test is actually quite
leaky, since any test after it will be executed with the mocked fetch.

This was the result of some rapid prototyping. I just want to see if the
idea has merit.